+++ module/system.js
@@ -113,7 +113,8 @@ Hooks.once("init", function () {
   foundry.applications.handlebars.loadTemplates([
     "systems/palladium-universal/templates/item/parts/effects.hbs",
     "systems/palladium-universal/templates/item/parts/description.hbs",
-    "systems/palladium-universal/templates/vehicle/time-machine.hbs"
+    "systems/palladium-universal/templates/vehicle/time-machine.hbs",
+    "systems/palladium-universal/templates/actor/parts/item-roll.hbs"
   ]);
 });